+++ cpp/src/remote_handle.cpp
@@ -19,8 +19,6 @@
 #include <cstring>
 #include <iostream>
 #include <memory>
-#include <optional>
-#include <regex>
 #include <sstream>
 #include <stdexcept>
 #include <string>
